% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_FuzzyNameMatch_function.R
\name{check_name_length_match}
\alias{check_name_length_match}
\title{A Levenshtein Sim Name List Function}
\usage{
check_name_length_match(input_name, search_name)
}
\arguments{
\item{string}{input and string list}
}
\value{
numeric output
}
\description{
This 'check_name_length_match' function applies function 'levenNameList' for
Where there isn't a match in the number of words, the Levenshtein distance calculation is iterated across each word permutation.
In both cases the mean of each result is returned.
}
\examples{
check_name_length_match(c("Derek", "Jones"), c("Derek", "Jones"))  ->  1
check_name_length_match(c("Derek", "Jones"), c("Derek", "Bones"))  ->   0.9
check_name_length_match(c("Rigobert", "Robert"), c("Derek", "Bones"))   ->   0.2291667
}
