% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_FuzzyNameMatch_function.R
\name{levelTwoListMatch}
\alias{levelTwoListMatch}
\title{A Levenshtein Sim Weighting and Selection Function to Return the Best Match}
\usage{
levelTwoListMatch(search_name_list, input_name_list)
}
\arguments{
\item{string}{input and string list}
}
\value{
numeric output
}
\description{
This 'levelTwoListMatch' function takes two lists of rider names and determines the positions of the best match.
Returns no match ('NA') where there is no good match or there are duplicates.
'search_name_list': The names (list) of the riders for whom we are seeking a match from another list of riders.
'input_name_list': The list containing names of riders for matching.
The position of the best match in the input_name_list of riders is returned.
}
