% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_FuzzyNameMatch_function.R
\name{levenBestMatch}
\alias{levenBestMatch}
\title{A Levenshtein Sim Wrapping Function to Return the Best Match}
\usage{
levenBestMatch(search_name, input_name_list)
}
\arguments{
\item{string}{input and string list}
}
\value{
numeric output
}
\description{
This 'levenBestMatch' function takes the list of Levenshtein values calculated for the entire list (of riders) and then determines which has the highest value (closest match) and then returns the position of the closest match.
'search_name': The name of the rider for whom we are seeking a match from another list of riders.
'input_name_list': The list containing names of riders.
The position of the best match in the input_name_list of riders is returned.
}
