% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_FuzzyNameMatch_function.R
\name{levenNameList}
\alias{levenNameList}
\title{A Levenshtein Sim value Function}
\usage{
levenNameList(input_word, input_name)
}
\arguments{
\item{string}{input and string list}
}
\value{
numeric output
}
\description{
Calculate the Levenshtein Sim value between an input word and each word in a name. This 'levenNameList' function is nested into the 'check_name_length_match' function.
}
\examples{
levenNameList("Derek", c("Derek", "Jones"))  ->  1
levenNameList("Frederik", c("Fred", "Backaert"))  ->   0.5
levenNameList("Rigoberto", c("Rigobert", "Robert"))   ->   0.8888889
}
