% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_tdfAuto_update_function.R
\name{tdfAuto_update}
\alias{tdfAuto_update}
\title{This function will update the documentation and install (locally) the latest
local changes to the tdfAuto package, using devtools.}
\usage{
tdfAuto_update()
}
\value{
Nothing to return
}
\description{
This function will update the documentation and install (locally) the latest
local changes to the tdfAuto package, using devtools.
}
