% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_text_clean_function.R
\name{text_clean}
\alias{text_clean}
\title{A Text Clean Function}
\usage{
text_clean(string)
}
\arguments{
\item{string}{input}
}
\value{
string output
}
\description{
Remove latin, foreign and special characters and replace them with standard western characters.
The intent is to have a single function for application uniformly across text import functions.
This function is primarily used to prepare data for writing to the MySQL database, which apparently doesn't accept a wide variety of foreign and special characters.
}
\examples{
text_clean("Yohann Gène")  ->  "Yohann Gene"
text_clean("Michał Périchon")   ->   "Michal Perichon"
}
